refactor: index.ts를 추가해 모듈 경로 간편화
