+++ src/components/error/index.ts
@@ -1,0 +1,7 @@
+import ErrorBox from './ErrorBox';
+import ErrorImg from './ErrorImg';
+import ErrorMessage from './ErrorMessage';
+import ErrorView from './ErrorView';
+import NullElementError from './NullElementError';
+
+export { ErrorBox, ErrorImg, ErrorView, ErrorMessage, NullElementError };
