+++ src/components/index.ts
@@ -1,0 +1,12 @@
+import Header from './Header';
+import Label from './Label';
+import MoreButton from './MoreButton';
+import RefreshButton from './RefreshButton';
+import SearchBox from './SearchBox';
+
+export * from './error';
+export * from './modal';
+export * from './movie';
+export * from './skeleton';
+
+export { Header, Label, MoreButton, RefreshButton, SearchBox };
