+++ src/components/modal/index.ts
@@ -1,0 +1,5 @@
+import AlertModal from './AlertModal';
+import ModalContainer from './ModalContainer';
+import ToastModal from './ToastModal';
+
+export { AlertModal, ModalContainer, ToastModal };
