+++ src/components/movie/MovieCard.ts
@@ -1,5 +1,5 @@
-import { Movie } from '../type/movie';
-import { createElementWithAttribute } from '../utils';
+import { Movie } from '../../type/movie';
+import { createElementWithAttribute } from '../../utils';
 
 import MovieImg from './MovieImg';
 import MovieScore from './MovieScore';
