+++ src/components/movie/index.ts
@@ -1,0 +1,21 @@
+import MovieCard from './MovieCard';
+import MovieImg from './MovieImg';
+import MovieItem from './MovieItem';
+import MovieList from './MovieList';
+import MovieListContainer from './MovieListContainer';
+import MovieListTitle from './MovieListTitle';
+import MovieScore from './MovieScore';
+import MovieTitle from './MovieTitle';
+import NoneMovieItem from './NoneMovieItem';
+
+export {
+  MovieCard,
+  MovieImg,
+  MovieItem,
+  MovieList,
+  MovieListContainer,
+  MovieListTitle,
+  MovieScore,
+  MovieTitle,
+  NoneMovieItem,
+};
